web/pyproject.toml
[tool.poetry]
name = "web"
version = "0.1.0"
description = ""
authors = ["uch <uch@energinet.dk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
dash = "^2.18.2"
pandas = "^2.2.3"
dash-mantine-components = "0.12.1"
pvlib = "^0.11.1"
diskcache = "^5.6.3"
scikit-learn = "^1.5.2"
numpy = "2.0.0"
adlfs = "^2024.7.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
